style: revert formatting changes, move Texture classes to auth/texture package
- Revert Javadoc alignment in GameProfile.java
- Revert comment line wrapping in ResolvableProfile.java
- Revert all line-length/indentation changes in MinecraftTypes.java
- Move Texture, TextureModel, TextureType to auth/texture/ package
- Update imports in GameProfile.java, MinecraftTypes.java, ResolvableProfile.java
