+++ protocol/src/main/java/org/geysermc/mcprotocollib/auth/texture/TextureModel.java
@@ -1,4 +1,4 @@
-package org.geysermc.mcprotocollib.auth;
+package org.geysermc.mcprotocollib.auth.texture;
 
 /**
  * The model used for a profile texture.
